% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract_timeline.R
\name{abstract_timeline}
\alias{abstract_timeline}
\title{Abstract the timeline}
\usage{
abstract_timeline(
  ftsList,
  sizing = list(columns = 2, tiles = 1, captions = 8, circles = 0.15, symbols = 2.5),
  stampCols = "darker",
  stampColsDif = 0.5,
  stampLoc = 1,
  timelineRows = NULL
)
}
\arguments{
\item{ftsList}{A list from \code{highlight_changes}}

\item{sizing}{A list of size specifications for the column names, the tiles, the captions, the cirlces, and the symbols}

\item{stampCols}{Either "darker" or "lighter" for stamp colour. Can enter a list corresponding to specific actions.}

\item{stampColsDif}{Degree to which stamp colour is darker or lighter. Can enter a list corresponding to specific actions.}

\item{stampLoc}{Location of stamp. 1 = top left. 2 = top right. 3 = bottom left. 4 = bottom right. 5 = center.}

\item{timelineRows}{Number of rows to divide the smallset timeline into.}
}
\description{
Abstract the timeline
}
