% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption_block.R
\name{add_caption_block}
\alias{add_caption_block}
\title{Add caption block}
\usage{
add_caption_block(i, col1, col2, col3, code)
}
\description{
A function to insert a caption block in the caption template
}
\keyword{internal}
