% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ghost_data.R
\name{add_ghost_data}
\alias{add_ghost_data}
\title{Add ghost data}
\usage{
add_ghost_data(itemNum, ghostDF1, ghostDF2, snapshotList)
}
\description{
A function to insert blank squares where data have been removed
}
\keyword{internal}
