% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plot.R
\name{build_plot}
\alias{build_plot}
\title{Build plot}
\usage{
build_plot(
  itemNum,
  accents,
  extTables,
  fourCols,
  ghostData,
  legendDF,
  maxDims,
  missingDataTints,
  output,
  printedData,
  sizing,
  smallsetTables,
  spacing,
  font,
  truncateData
)
}
\description{
Builds snapshot and resume marker plots.
}
\keyword{internal}
