% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeline.R
\name{create_timeline}
\alias{create_timeline}
\title{Create the timeline}
\usage{
create_timeline(
  snapshotList,
  constant = list("#cecfd6", 0.7),
  changed = list("#0f3d1c", 0.7),
  added = list("#a35222", 0.7),
  deleted = list("#3e4d63", 0.7),
  abstract = TRUE,
  ghostData = FALSE,
  highlightNA = FALSE,
  sizing = list(columns = 2, tiles = 1, captions = 8, circles = 0.15, symbols = 2.5,
    data = 2.5, legend = 7),
  truncateData = FALSE,
  accentCols = "darker",
  accentColsDif = 0.5,
  otherTextCol = 1,
  stampLoc = 1,
  timelineRows = NULL,
  timelineFont = "sans",
  captionSpace = 1,
  captionTemplateName = NULL,
  captionTemplateDir = NULL
)
}
\arguments{
\item{snapshotList}{A list from \code{highlight_changes}}

\item{constant}{A colour to represent data that have not changed.}

\item{changed}{A colour to represent data that have changed.}

\item{added}{A colour to represent data that have been added.}

\item{deleted}{A colour to represent data that have been deleted.}

\item{abstract}{TRUE or FALSE for hiding data values in timeline.}

\item{ghostData}{TRUE or FALSE for including blank tiles where data have been removed.}

\item{highlightNA}{TRUE or FALSE for using a lighter colour to signal data value is missing.}

\item{sizing}{A list of size specifications for the column names, the tiles, the captions, the circles, and the symbols}

\item{truncateData}{FALSE if data do not need to be truncated to fit in tiles. Otherwise, an integer specifying width of data value (width includes "...").}

\item{accentCols}{Either "darker" or "lighter" for stamp colour. Can enter a list corresponding to specific actions.}

\item{accentColsDif}{Degree to which stamp colour is darker or lighter. Can enter a list corresponding to specific actions.}

\item{otherTextCol}{Value between 0 and 1. Default is 1, meaning column names will be black. 0 means columns will be the same colour as the constant colour.}

\item{stampLoc}{Location of stamp. 1 = top left. 2 = top right. 3 = bottom left. 4 = bottom right. 5 = center.}

\item{timelineRows}{Number of rows to divide the smallset timeline into.}

\item{timelineFont}{Font family.}

\item{captionSpace}{Increase room for captions. Any value greater than or equal to .5. Default is one.}

\item{captionTemplateName}{Name of caption template if renamed so not written over when running highlight_changes function.}

\item{captionTemplateDir}{Name of caption template directory if moved so not written over when running highlight_changes function.}
}
\description{
Create the timeline
}
