% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract snapshot tables}
\usage{
extract_tables(itemNum, snapshotList)
}
\description{
The function extracts data and colour information from the
  flextable objects.
}
\keyword{internal}
