% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_plot.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Finalise plot}
\usage{
finalise_plot(itemNum, plots, output, maxDims, font, sizing, spacing)
}
\description{
Adds captions and standard dimensions to the plots.
}
\keyword{internal}
