% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data_changes.R
\name{find_data_changes}
\alias{find_data_changes}
\title{Find data changes}
\usage{
find_data_changes(smallsetList, fourCols, altText)
}
\description{
Finds data edits, additions, and deletions between snapshots.
}
\keyword{internal}
