% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_to_colour_plot.R
\name{flex_to_colour_plot}
\alias{flex_to_colour_plot}
\title{Flex to colour plot}
\usage{
flex_to_colour_plot(ftsItemNum, ftsList, sizing)
}
\description{
A function to transform the flextable into a colour plot
}
\keyword{internal}
