% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alt_text.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate Alt Text}
\usage{
generate_alt_text(snapshotList, altTextInfo, l, abstract, ghostData)
}
\description{
The function produces alt text for the Smallset Timeline figure.
}
\keyword{internal}
