% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeline_dimensions.R
\name{get_timeline_dimensions}
\alias{get_timeline_dimensions}
\title{Get timeline dimensions}
\usage{
get_timeline_dimensions(tabs)
}
\description{
Finds the maximum and minimum number of rows and columns in the Smallset Timeline.
}
\keyword{internal}
