% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_timeline_plot2.R
\name{make_timeline_plot2}
\alias{make_timeline_plot2}
\title{Make a timeline plot}
\usage{
make_timeline_plot2(
  itemNum,
  snapshotList,
  abstract,
  sizing,
  truncateData,
  accentCols,
  accentColsDif,
  otherTextCol,
  stampLoc,
  maxDims,
  timelineFont,
  captionSpace,
  accents,
  legendDF,
  ghostDF1,
  ghostDF2,
  highlightNA,
  captionTemplateName,
  captionTemplateDir
)
}
\description{
A function to transform the flextable into a colour plot
}
\keyword{internal}
