% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Synethic dataset.}
\format{
A data frame with 100 rows and 8 variables:
\describe{
  \item{C1}{categorical}
  \item{C2}{binary}
  \item{C3}{discrete}
  \item{C4}{discrete}
  \item{C5}{continuous}
  \item{C6}{continuous}
  \item{C7}{continuous}
  \item{C8}{continuous}
}
}
\usage{
mydata
}
\description{
A synthetic dataset generated for illustrative purposes.
}
\keyword{datasets}
