% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_colour_sheet.R
\name{prepare_colour_sheet}
\alias{prepare_colour_sheet}
\title{Prepare colour sheet}
\usage{
prepare_colour_sheet(smallsetList, fourCols)
}
\description{
Prepares the visual appearance matrix for automated Smallset
  selection.
}
\keyword{internal}
