% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_score_sheet.R
\name{prepare_score_sheet}
\alias{prepare_score_sheet}
\title{Prepare score sheet}
\usage{
prepare_score_sheet(smallsetList = smallsetList, data = data)
}
\description{
Generates a score sheet for automatic Smallset selection
  between snapshots.
}
\keyword{internal}
