% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_smallset.R
\name{prepare_smallset}
\alias{prepare_smallset}
\title{Prepare Smallset}
\usage{
prepare_smallset(data, code, dir, rowCount, rowNums, auto, ignoreCols)
}
\description{
Prepares the Smallset.
}
\keyword{internal}
