% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_dimensions.R
\name{retrieve_dimensions}
\alias{retrieve_dimensions}
\title{Retrieve dimensions}
\usage{
retrieve_dimensions(tabs)
}
\description{
Retrieves the maximum and minimum number of rows and columns in
  the Smallset Timeline.
}
\keyword{internal}
