% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_tables.R
\name{retrieve_tables}
\alias{retrieve_tables}
\title{Retrieve snapshot tables}
\usage{
retrieve_tables(itemNum, smallsetTables)
}
\description{
Retrieves data and colour information from the snapshot tables.
}
\keyword{internal}
