% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_scheme.R
\name{return_scheme}
\alias{return_scheme}
\title{Return selected colour scheme}
\usage{
return_scheme(colScheme = "colScheme1")
}
\description{
A function to return the selected colour scheme.
}
\keyword{internal}
