% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simple_gurobi.R
\name{run_simple_gurobi}
\alias{run_simple_gurobi}
\title{Run simple gurobi}
\usage{
run_simple_gurobi(data = data, code = code, dir = dir, rowCount = rowCount)
}
\description{
Runs simple gurobi selection model.
}
\keyword{internal}
