% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_data.R
\docType{data}
\name{s_data}
\alias{s_data}
\title{Synthetic dataset}
\format{
A data frame with 100 rows and 8 columns:
\describe{
  \item{C1}{discrete}
  \item{C2}{binary}
  \item{C3}{discrete}
  \item{C4}{discrete}
  \item{C5}{continuous}
  \item{C6}{continuous}
  \item{C7}{continuous}
  \item{C8}{continuous}
}
}
\usage{
s_data
}
\description{
A synthetic dataset generated for illustrative purposes.
}
\examples{
str(s_data)
}
\keyword{datasets}
