% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_smallset.R
\name{select_smallset}
\alias{select_smallset}
\title{Select smallset}
\usage{
select_smallset(
  data,
  rowCount = 6,
  rowNums = NULL,
  runBig = FALSE,
  ignoreCols = ignoreCols
)
}
\description{
The function selects the rows included in the smallset.
}
\keyword{internal}
