% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets_labelling.R
\name{sets_labelling}
\alias{sets_labelling}
\title{Sets labelling}
\usage{
sets_labelling(labelCol = NULL, labelColDif = NULL)
}
\arguments{
\item{labelCol}{"lighter" or "darker" for column names
and printed data, in comparison to tile colours.}

\item{labelColDif}{Value between 0-1 determining how much lighter or
darker.}
}
\description{
Sets labelling parameters for the Smallset Timeline.
}
\examples{
# labels are black
sets_labelling(labelCol = "darker", labelColDif = 1)

# labels are white
sets_labelling(labelCol = "lighter", labelColDif = 1)

# midpoint between tile colours and black
sets_labelling(labelCol = "darker", labelColDif = .5)

}
