% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_smallset_colours.R
\name{view_schemes}
\alias{view_schemes}
\title{View schemes}
\usage{
view_schemes(scheme = "all")
}
\arguments{
\item{scheme}{One of "colScheme1", "colScheme2", "colScheme3", or "all".}
}
\description{
This command is for viewing built-in colour schemes.
Schemes can be viewed together or individually.
}
