% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_caption_template.R
\name{write_caption_template}
\alias{write_caption_template}
\title{Write caption template}
\usage{
write_caption_template(authorName, script, pathway)
}
\description{
A function to prepare the caption template
}
\keyword{internal}
