% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text_templates.R
\name{write_colourUnchanged}
\alias{write_colourUnchanged}
\title{Write unchanged colour}
\usage{
write_colourUnchanged(colourUnchanged)
}
\description{
Writes about the unchanged colour.
}
\keyword{internal}
