% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_text_templates.R
\name{write_columnsDelete}
\alias{write_columnsDelete}
\title{Write column deletions}
\usage{
write_columnsDelete(columnsDelete, colourDelete)
}
\description{
Writes about column deletions.
}
\keyword{internal}
