% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_smallset_code.R
\name{write_smallset_code}
\alias{write_smallset_code}
\title{Write smallset code}
\usage{
write_smallset_code(
  scriptName,
  dir,
  ignoreCols,
  keepCols,
  smallset,
  lang,
  modelSelection = FALSE
)
}
\description{
The function turns the preprocessing code into a function that
  takes snapshots.
}
\keyword{internal}
