% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract_timeline.R
\name{abstract_timeline}
\alias{abstract_timeline}
\title{Abstract the timeline}
\usage{
abstract_timeline(
  ftsList,
  sizing = list(columns = 2, tiles = 1, captions = 8, circles = 0.15, symbols = 2.5)
)
}
\arguments{
\item{ftsList}{A list from \code{highlight_changes}}

\item{sizing}{A list of size specifications for the column names, the tiles, the captions, the cirlces, and the symbols}
}
\description{
Abstract the timeline
}
