% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeline.R
\name{create_timeline}
\alias{create_timeline}
\title{Create the Timeline}
\usage{
create_timeline(
  snapshotList,
  colours,
  printedData,
  ghostData,
  missingDataTints,
  sizing,
  truncateData,
  rotateHeader,
  headerSpace,
  accentCol,
  accentColDif,
  otherTextCol,
  timelineRows,
  timelineFont,
  captionSpace
)
}
\description{
Creates the Timeline.
}
\keyword{internal}
