% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeline.R
\name{create_timeline}
\alias{create_timeline}
\title{Create the timeline}
\usage{
create_timeline(
  ftsList,
  abstract = TRUE,
  sizing = list(columns = 2, tiles = 1, captions = 8, circles = 0.15, symbols = 2.5,
    data = 2.5),
  accentCols = "darker",
  accentColsDif = 0.5,
  stampLoc = 1,
  timelineRows = NULL,
  timelineFont = "sans"
)
}
\arguments{
\item{ftsList}{A list from \code{highlight_changes}}

\item{abstract}{TRUE or FALSE for hiding data values in timeline.}

\item{sizing}{A list of size specifications for the column names, the tiles, the captions, the cirlces, and the symbols}

\item{accentCols}{Either "darker" or "lighter" for stamp colour. Can enter a list corresponding to specific actions.}

\item{accentColsDif}{Degree to which stamp colour is darker or lighter. Can enter a list corresponding to specific actions.}

\item{stampLoc}{Location of stamp. 1 = top left. 2 = top right. 3 = bottom left. 4 = bottom right. 5 = center.}

\item{timelineRows}{Number of rows to divide the smallset timeline into.}

\item{timelineFont}{Font family.}
}
\description{
Create the timeline
}
