% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeline.R
\name{create_timeline}
\alias{create_timeline}
\title{Create the timeline}
\usage{
create_timeline(
  snapshotList,
  abstract = TRUE,
  sizing = list(columns = 2, tiles = 1, captions = 8, circles = 0.15, symbols = 2.5,
    data = 2.5, legend = 7),
  truncateData = FALSE,
  accentCols = "darker",
  accentColsDif = 0.5,
  otherTextCol = 1,
  stampLoc = 1,
  timelineRows = NULL,
  timelineFont = "sans",
  captionSpace = 1
)
}
\arguments{
\item{snapshotList}{A list from \code{highlight_changes}}

\item{abstract}{TRUE or FALSE for hiding data values in timeline.}

\item{sizing}{A list of size specifications for the column names, the tiles, the captions, the cirlces, and the symbols}

\item{truncateData}{FALSE if data do not need to be truncated to fit in tiles. Otherwise, an integer specifying width of data value (width includes "...").}

\item{accentCols}{Either "darker" or "lighter" for stamp colour. Can enter a list corresponding to specific actions.}

\item{accentColsDif}{Degree to which stamp colour is darker or lighter. Can enter a list corresponding to specific actions.}

\item{otherTextCol}{Value between 0 and 1. Default is 1, meaning column names will be black. 0 means columns will be the same colour as the constant colour.}

\item{stampLoc}{Location of stamp. 1 = top left. 2 = top right. 3 = bottom left. 4 = bottom right. 5 = center.}

\item{timelineRows}{Number of rows to divide the smallset timeline into.}

\item{timelineFont}{Font family.}

\item{captionSpace}{Increase room for captions. Any value greater than or equal to .5. Default is one.}
}
\description{
Create the timeline
}
