% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alt_text.R
\name{generate_alt_text}
\alias{generate_alt_text}
\title{Generate alt text}
\usage{
generate_alt_text(
  tables,
  fourCols,
  legendDF,
  atInfo,
  l,
  printedData,
  ghostData
)
}
\description{
Generates alternative text (alt text) for the Smallset Timeline.
}
\keyword{internal}
