% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_changes.R
\name{highlight_changes}
\alias{highlight_changes}
\title{Highlight changes}
\usage{
highlight_changes(
  smallsetList = smallsetList,
  tempName = captionTemplateName,
  tempDir = captionTemplateDir,
  tempAuthor = captionTemplateAuthor
)
}
\description{
A function to identify changes, additions, and deletions between snapshots
}
\keyword{internal}
