% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_timeline_plot.R
\name{make_timeline_plot}
\alias{make_timeline_plot}
\title{Make a timeline plot}
\usage{
make_timeline_plot(
  ftsItemNum,
  ftsList,
  abstract,
  sizing,
  accentCols,
  accentColsDif,
  stampLoc,
  maxDims,
  timelineFont,
  accents
)
}
\description{
A function to transform the flextable into a colour plot
}
\keyword{internal}
