% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_timeline_plot.R
\name{make_timeline_plot}
\alias{make_timeline_plot}
\title{Make timeline plot}
\usage{
make_timeline_plot(
  itemNum,
  extTables,
  snapshotList,
  printedData,
  ghostData,
  sizing,
  truncateData,
  rotateHeader,
  headerSpace,
  accentCols,
  accentColsDif,
  otherTextCol,
  maxDims,
  timelineFont,
  captionSpace,
  accents,
  legendDF,
  missingDataTints,
  timelineRows
)
}
\description{
The function transforms a table snapshot into a plot snapshot.
}
\keyword{internal}
