% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snapshots.R
\name{plot_snapshots}
\alias{plot_snapshots}
\title{Plot the snapshots}
\usage{
plot_snapshots(
  itemNum,
  accents,
  extTables,
  fourCols,
  ghostData,
  legendDF,
  maxDims,
  missingDataTints,
  output,
  printedData,
  sizing,
  smallsetTables,
  spacing,
  font,
  truncateData
)
}
\description{
Transforms a table snapshot into a plot snapshot.
}
\keyword{internal}
