% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_smallset.R
\name{prepare_smallset}
\alias{prepare_smallset}
\title{Prepare smallset}
\usage{
prepare_smallset(
  data,
  code,
  dir = getwd(),
  rowCount = 6,
  rowNums = NULL,
  runBig = FALSE,
  ignoreCols = NULL,
  captionTemplateName = "captionTemplate",
  captionTemplateDir = getwd(),
  captionTemplateAuthor = NULL
)
}
\arguments{
\item{data}{Data set.}

\item{code}{Script with data preprocessing code for data set.}

\item{dir}{File path to data preprocessing code.}

\item{rowCount}{Integer greater than or equal to 5. Number of rows to include
in the smallset.}

\item{rowNums}{Numeric vector of row numbers. Indicates particular rows from
the data set to be included in the smallset.}

\item{runBig}{TRUE or FALSE. FALSE means preprocessing code will be run on
smallset. TRUE means preprocessing code will be run on the big data set,
and the smallset will be extracted from that output at each snap point.}

\item{ignoreCols}{Character vector of column names. Indicates which columns
from the data set should not be included in the smallset. Columns in this
vector should usually not be referenced in the data preprocessing code.}

\item{captionTemplateName}{File name for the caption template.}

\item{captionTemplateDir}{File path for the caption template.}

\item{captionTemplateAuthor}{Name of author for the caption template.}
}
\description{
The function selects a smallset, takes snapshots, and
  identifies the changes between snapshots. The output can be passed to
  \code{create_timeline} to create a smallset timeline.
}
