% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_smallset.R
\name{prepare_smallset}
\alias{prepare_smallset}
\title{Prepare smallset}
\usage{
prepare_smallset(data, code, rowCount = 6, rowNums = NULL, dir = getwd())
}
\arguments{
\item{data}{A data set}

\item{code}{A script to prepare data for use in modelling}

\item{rowCount}{The number of rows to include in the smallset}

\item{rowNums}{A numeric vector of row numbers indicating particular rows from the data set to be included in the smallset}

\item{dir}{A file path for the smallset script}
}
\description{
Prepare smallset
}
