% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_smallset.R
\name{prepare_smallset}
\alias{prepare_smallset}
\title{Prepare smallset}
\usage{
prepare_smallset(
  data,
  code,
  dir = getwd(),
  rowCount = 6,
  rowNums = NULL,
  runBig = FALSE,
  ignoreCols = NULL,
  captionTemplateName = "captionTemplate",
  captionTemplateDir = getwd(),
  captionTemplateAuthor = NULL
)
}
\arguments{
\item{data}{A data set}

\item{code}{A script to prepare data for use in modelling}

\item{dir}{A file path for the smallset script}

\item{rowCount}{The number of rows to include in the smallset}

\item{rowNums}{A numeric vector of row numbers indicating particular rows from the data set to be included in the smallset}

\item{runBig}{Default is FALSE (preprocessing code is run on smallset) - if TRUE, preprocessing code is run on the original dataset and smallset snapshots are extracted from that}

\item{ignoreCols}{A vector of column names that you do not want included in the timeline (cannot be referenced in preprocessing code)}

\item{captionTemplateName}{A file name for the caption .Rmd template}

\item{captionTemplateDir}{A file path for the caption .Rmd template}

\item{captionTemplateAuthor}{The author's name for the caption .Rmd file}
}
\description{
Prepare smallset
}
