% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_captions_rmd.R
\name{read_captions_rmd}
\alias{read_captions_rmd}
\title{Read the caption R markdown file}
\usage{
read_captions_rmd(script, pathway)
}
\description{
A function to interpret the caption R markdown file
}
\keyword{internal}
