% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_advanced_gurobi.R
\name{run_advanced_gurobi}
\alias{run_advanced_gurobi}
\title{Run advanced gurobi}
\usage{
run_advanced_gurobi(data, code, rowCount, lang, fourCols)
}
\description{
Runs the coverage+variety optimisation problem.
}
\keyword{internal}
