% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_smallset.R
\name{select_smallset}
\alias{select_smallset}
\title{Select smallset}
\usage{
select_smallset(dataNum, data, rowCount = 6, rowNums = NULL)
}
\description{
A function to prepare a smallset
}
\keyword{internal}
