% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smallset_Timeline.R
\name{Smallset_Timeline}
\alias{Smallset_Timeline}
\title{Smallset Timeline}
\usage{
Smallset_Timeline(
  data,
  code,
  dir = getwd(),
  rowCount = 6,
  rowNums = NULL,
  auto = NULL,
  ignoreCols = NULL,
  colours = "colScheme1",
  printedData = FALSE,
  ghostData = TRUE,
  missingDataTints = FALSE,
  sizing = list(columns = 3, captions = 3, tiles = 0.1, data = 2.5, legendText = 10,
    legendIcons = 1, resume = 0.25),
  truncateData = FALSE,
  rotateHeader = FALSE,
  headerSpace = c(1, 0.5),
  accentCol = "darker",
  accentColDif = 0.8,
  otherTextCol = 1,
  timelineRows = 1,
  timelineFont = "sans",
  captionSpace = 3
)
}
\arguments{
\item{data}{Dataset.}

\item{code}{R or Python script with data preprocessing code for the dataset.
Filename extension should be included (e.g., "my_code.R" or "my_code.py").}

\item{dir}{File path to the data preprocessing code. Default is the working directory.}

\item{rowCount}{Integer greater than or equal to 5. Number of rows to include
in the smallset.}

\item{rowNums}{Numeric vector of row numbers. Indicates particular rows from
the data set to be included in the smallset.}

\item{auto}{1 or 2. 1 = simple gurobi model selection. 2 = advanced gurobi
model selction.}

\item{ignoreCols}{Character vector of column names. Indicates which columns
from the data set should not be included in the smallset. Columns in this
vector should usually not be referenced in the data preprocessing code.}

\item{colours}{how to select colours}

\item{printedData}{TRUE or FALSE. TRUE prints data values in tables.}

\item{ghostData}{TRUE or FALSE. TRUE includes blank spaces where data have
been removed.}

\item{missingDataTints}{TRUE or FALSE. TRUE plots a lighter colour value to signal
data value is missing.}

\item{sizing}{List of size specifications.}

\item{truncateData}{TRUE or FALSE. FALSE if data do not need to be truncated
to fit within table tiles. Otherwise, an integer specifying width of data
value (width includes "...").}

\item{rotateHeader}{TRUE or FALSE. If TRUE, column names are printed at 45
degree angle. Use if column names overlap when set to FALSE.}

\item{headerSpace}{Vector. Default is c(1, .5). First value corresponds to
room above the table and second to the right of the table.}

\item{accentCol}{Either "darker" or "lighter."}

\item{accentColDif}{Value between 0 and 1. Corresponds to how much lighter
or darker accent colour will be.}

\item{otherTextCol}{Value between 0 and 1. Default is 1, which is when column
names are black. 0 means columns will be the constant colour.}

\item{timelineRows}{Integer greater than or equal to one. Number of rows to
divide the smallset timeline into.}

\item{timelineFont}{One of "sans", "serif", or "mono".}

\item{captionSpace}{Value greater than or equal to .5. Higher values create
more caption space. Default is 1.}
}
\description{
The command for creating a Smallset Timeline.
}
