% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smallset_Timeline.R
\name{Smallset_Timeline}
\alias{Smallset_Timeline}
\title{Smallset Timeline}
\usage{
Smallset_Timeline(
  data,
  code,
  rowCount = 5,
  rowSelect = NULL,
  rowReturn = FALSE,
  rowNums = NULL,
  ignoreCols = NULL,
  colours = 1,
  altText = FALSE,
  printedData = FALSE,
  truncateData = NULL,
  ghostData = TRUE,
  missingDataTints = FALSE,
  font = "sans",
  sizing = list(columns = 2.5, captions = 2.5, tiles = 0.2, data = 2.5, legendText =
    10, legendIcons = 1, resume = 0.25),
  spacing = list(captionB = 2, columnsT = 1, tablesR = 0.5, rows = 1, columnsDeg = 0),
  labelling = list(labelsCol = "darker", labelsColDif = 0.5)
)
}
\arguments{
\item{data}{Dataset that is being preprocessed.}

\item{code}{R or Python data preprocessing script. Include the filename extension 
(e.g., "my_code.R" or "my_code.py"). If the script is not located in the working 
directory, include the file path.}

\item{rowCount}{Integer between 5-15 for number of Smallset rows.}

\item{rowSelect}{NULL, 1, or 2. When NULL, Smallset rows are randomly sampled.
When 1, Smallset rows are selected using the coverage optimisation model. 
When 2, Smallset rows are selected using the coverage + variety optimisation model, 
which has a long run time for large datasets. Options 1 and 2 use the
Gurobi solver and require a Gurobi license. Please visit https://www.gurobi.com to obtain 
a license (free academic licenses are available).}

\item{rowReturn}{A logical. TRUE prints, to the console, the row numbers 
of the rows selected for the Smallset.}

\item{rowNums}{Numeric vector indicating particular rows from the dataset 
to include in the Smallset.}

\item{ignoreCols}{Character vector of column names indicating which to exclude 
from the Smallset. These columns can't be referenced in the data preprocessing code.}

\item{colours}{Either 1, 2, or 3 for one of the pre-built colour schemes (all are 
colourblind friendly and 3 is black/white printer friendly) or a list with 
four hex colour codes for same, edit, add, and delete (e.g., list(same = "#E6E3DF", 
edit = "#FFC500", add = "#5BA2A6", delete = "#DDC492")).}

\item{altText}{A logical. TRUE generates alternative text (alt text)
for the Smallset Timeline and prints it to the console.}

\item{printedData}{A logical. TRUE prints data values in the Smallset
snapshots.}

\item{truncateData}{Integer specifying the number of characters to print for each
data value (results in characters + "..."). Default is NULL, where entire data 
value is printed.}

\item{ghostData}{A logical. TRUE includes blank spaces where data have
been removed.}

\item{missingDataTints}{A logical. TRUE plots a lighter colour value to
signal a missing data value.}

\item{font}{Any font you have installed in R. Default is "sans".}

\item{sizing}{List of size specifications: column names (columns), caption
text (captions), tile size, (tiles), printed data (data), legend text
(legendText), legend icons (legendIcons), and resume markers (resume). List
of defaults: list(columns = 2.5, captions = 2.5, tiles = .2, data = 2.5,
legendText = 10, legendIcons = 1, resume = .25).}

\item{spacing}{List of spacing specifications: space below captions
(captionB), space above columns (columnsT), space to the right of Smallset
tables (tablesR), number of Timeline rows (rows), and degree rotation of
column names (columnsDeg). List of defaults: list(captionB = 2, columnsT =
1, tablesR = .5, rows = 1, columnsDeg = 0).}

\item{labelling}{List of colour specifications for labelling, which includes
column names and printed data. First specification is whether the labels
should be lighter or darker than the tile colours (labelsCol). Second
specification is a value between 0 and 1 for how much lighter or darker it
should be (labelsColDif). Example 1: list(labelsCol = "darker", labelsColDif
= 1) means labels are black. Example 2: list(labelsCol = "lighter",
labelsColDif = 1) means labels are white. Default is list(labelsCol =
"darker", labelsColDif = .5), which is midpoint between tile colours and
black.}
}
\value{
A plot.
}
\description{
Creates a Smallset Timeline to visualise data preprocessing decisions.
}
\details{
Prior to running this command, you will need to add structured
 comments to your R or Python data preprocessing script, providing snapshot
 points and captions. \itemize{\item{`# smallsets start mydata` - start
 tracking code and take the first data snapshot, where "mydata" is the name
 of your data object} \item{`# smallsets snap mydata` - take a data snapshot after the
 next line of code} \item{`# smallsets end mydata` - stop tracking code and
 take the last data snapshot} \item{Snapshot captions are added between
 caption brackets, `caption[...]caption`, at the end of the comments listed
 above}}
}
\examples{
set.seed(145)

Smallset_Timeline(
  data = mydata,
  code = system.file("preprocess_data.R", package = "smallsets")
)
}
