% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smallset_Timeline.R
\name{Smallset_Timeline}
\alias{Smallset_Timeline}
\title{Smallset Timeline}
\usage{
Smallset_Timeline(
  data,
  code,
  dir = getwd(),
  rowCount = 6,
  rowNums = NULL,
  autoSelect = NULL,
  ignoreCols = NULL,
  colours = 1,
  altText = FALSE,
  printedData = FALSE,
  truncateData = NULL,
  ghostData = TRUE,
  missingDataTints = FALSE,
  timelineFont = "sans",
  sizing = list(columns = 3, captions = 3, tiles = 0.1, data = 2.5, legendText = 10,
    legendIcons = 1, resume = 0.25),
  spacing = list(captionB = 3, columnsT = 1, tablesR = 0.5, rows = 1, columnsDeg = 0),
  labelling = list(labelsCol = "darker", labelsColDif = 0.5)
)
}
\arguments{
\item{data}{The dataset that is being preprocessed.}

\item{code}{The R or Python data preprocessing script. Filename extension
should be included (e.g., "my_code.R" or "my_code.py").}

\item{dir}{File path to the data preprocessing code. Default is the working
directory.}

\item{rowCount}{Integer between 5-15 for number of rows in the Smallset.}

\item{rowNums}{Numeric vector indicating particular rows from the dataset 
to include in the Smallset.}

\item{autoSelect}{1, 2, or NULL (default). 1 = select Smallset rows using the
coverage optimisation model. 2 = select Smallset rows using the
coverage+variety optimisation model (can take a long time to run). These
optimisation problems are solved using Gurobi. Please visit
https://www.gurobi.com to obtain a gurobi license (free academic licenses
are available) in order to use these selection methods. Otherwise, leave
autoSelect = NULL, and the rows will be randomly sampled from the dataset.}

\item{ignoreCols}{Character vector of column names indicating which to exclude 
from the Smallset. These columns can't be referenced in the data preprocessing code.}

\item{colours}{Either one of three pre-built colour schemes (1, 2, or 3 - all are 
colourblind friendly and 3 is black and white printer friendly) or a list with 
four hex colour codes for
 same, edit, add, and delete (e.g., list(same = "#E6E3DF", edit = "#FFC500",
 add = "#5BA2A6", delete = "#DDC492")).}

\item{altText}{A logical. TRUE generates alternative text (alt text)
for the Smallset Timeline and prints it to the console.}

\item{printedData}{A logical. TRUE prints data values in the Smallset
snapshots.}

\item{truncateData}{Integer specifying the number of characters to print for each
data value (results in characters + "..."). Default is NULL, where entire data 
value is printed.}

\item{ghostData}{A logical. TRUE includes blank spaces where data have
been removed.}

\item{missingDataTints}{A logical. TRUE plots a lighter colour value to
signal a missing data value.}

\item{timelineFont}{Any font you have installed in R. Default is "sans".}

\item{sizing}{List of size specifications: column names (columns), caption
text (captions), tile size, (tiles), printed data (data), legend text
(legendText), legend icons (legendIcons), and resume markers (resume). List
of defaults: list(columns = 3, captions = 3, tiles = .1, data = 2.5,
legendText = 10, legendIcons = 1, resume = .25).}

\item{spacing}{List of spacing specifications: space below captions
(captionB), space above columns (columnsT), space to the right of Smallset
tables (tablesR), number of Timeline rows (rows), and degree rotation of
column names (columnsDeg). List of defaults: list(captionB = 3, columnsT =
1, tablesR = .5, rows = 1, columnsDeg = 0).}

\item{labelling}{List of colour specifications for labelling, which includes
column names and printed data. First specification is whether the labels
should be lighter or darker than the tile colours (labelsCol). Second
specification is a value between 0 and 1 for how much lighter or darker it
should be (labelsColDif). Example 1: list(labelsCol = "darker", labelsColDif
= 1) means labels are black. Example 2: list(labelsCol = "lighter",
labelsColDif = 1) means labels are white. Default is list(labelsCol =
"darker", labelsColDif = .5), which is midpoint between tile colours and
black.}
}
\value{
A plot.
}
\description{
This is the command for creating a Smallset Timeline to visualise
 data preprocessing decisions.
}
\details{
Prior to running this command, you will need to add structured
 comments to your R or Python data preprocessing script, providing snapshot
 points and captions. \itemize{\item{`# start smallset mydata` - start
 tracking code and take the first data snapshot, where "mydata" is the name
 of your data object} \item{`# snap mydata` - take a data snapshot after the
 next line of code} \item{`# end smallset mydata` - stop tracking code and
 take the last data snapshot} \item{Snapshot captions are added between
 caption brackets, `caption[...]caption`, at the end of the comments listed
 above}}
}
\examples{
set.seed(145)

Smallset_Timeline(
  data = mydata,
  code = system.file("preprocess_data.R", package = "smallsets")
)
}
