% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_smallset_colours.R
\name{view_smallset_colours}
\alias{view_smallset_colours}
\title{View smallset colour schemes}
\usage{
view_smallset_colours(scheme = "all")
}
\arguments{
\item{scheme}{Either "colScheme1", "colScheme2", "colScheme3", or "all".}
}
\description{
This function is for viewing a built-in smallset colour scheme. Schemes can be viewed together or individually.
}
