% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_caption_template.R
\name{write_caption_template}
\alias{write_caption_template}
\title{Write caption template}
\usage{
write_caption_template(authorName, script, pathway, lang)
}
\description{
This function creates a caption template (Rmd file).
}
\keyword{internal}
