% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_smallset_code.R
\name{write_smallset_code}
\alias{write_smallset_code}
\title{Write Smallset code}
\usage{
write_smallset_code(code, smallset, lang)
}
\description{
Turns the preprocessing code into a function that takes
  snapshots.
}
\keyword{internal}
